+++ src/celestia/win32/winlocations.cpp
@@ -19,7 +19,7 @@ static const int FeatureSizeSliderRange = 100;
 static const float MinFeatureSize = 1.0f;
 static const float MaxFeatureSize = 100.0f;
 
-static const uint32 FilterOther = ~(Location::City |
+static const uint64_t FilterOther = ~(Location::City |
                                     Location::Observatory |
                                     Location::LandingSite |
                                     Location::Crater |
@@ -60,7 +60,7 @@ static BOOL APIENTRY LocationsProc(HWND hDlg,
     case WM_COMMAND:
     {
         Observer* obs = dlg->appCore->getSimulation()->getActiveObserver();
-        uint32 locationFilter = obs->getLocationFilter();
+        uint64_t locationFilter = obs->getLocationFilter();
 
         switch (LOWORD(wParam))
         {
@@ -172,7 +172,7 @@ LocationsDialog::LocationsDialog(HINSTANCE appInstance,
 }
 
 
-static void dlgCheck(HWND hDlg, WORD item, uint32 flags, uint32 f)
+static void dlgCheck(HWND hDlg, WORD item, uint64_t flags, uint64_t f)
 {
     SendDlgItemMessage(hDlg, item, BM_SETCHECK,
                        ((flags & f) != 0) ? BST_CHECKED : BST_UNCHECKED, 0);
@@ -182,7 +182,7 @@ static void dlgCheck(HWND hDlg, WORD item, uint32 flags, uint32 f)
 void LocationsDialog::SetControls(HWND hDlg)
 {
     Observer* obs = appCore->getSimulation()->getActiveObserver();
-    uint32 locFilter = obs->getLocationFilter();
+    uint64_t locFilter = obs->getLocationFilter();
 
     dlgCheck(hDlg, IDC_SHOW_CITIES,        locFilter, Location::City);
     dlgCheck(hDlg, IDC_SHOW_OBSERVATORIES, locFilter, Location::Observatory);
